mi_proyecto/requirements.txt
Django==5.1.4
django-allauth==65.3.0
django-extensions==3.2.3
djangorestframework==3.15.2
pillow==11.0.0